pyproject.toml
[tool.poetry]
name = "bs"
version = "0.1.0"
description = "Bs API"
authors = ["Samuel Ivanaskas <samuelivanaskas@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.0"
factory-boy = "^3.3.0"
django = "^4.2.4"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
